requirements.txt
fastapi==0.116.1
pyjwt==2.10.1
uvicorn==0.35.0
pytest==8.4.1
httpx==0.28.1